+++ internal/app/shell.go
@@ -259,6 +259,7 @@ func Shell(props ShellProps) uic.Node {
 		uic.CreateElement(TxnLinkHost),
 		uic.CreateElement(AccountEditHost),
 		uic.CreateElement(BudgetEditHost),
+		uic.CreateElement(AutoBudgetHost),
 		uic.CreateElement(GoalEditHost),
 		uic.CreateElement(TaskEditHost),
 		uic.CreateElement(MemberEditHost),